Changed home position coordinates being int type

+++ src/main/java/io/fleetcoreplatform/Models/DroneIdentity.java
@@ -1,5 +1,5 @@
 package io.fleetcoreplatform.Models;
 
 public record DroneIdentity(String name, Home home) {
-    public record Home(int x, int y, int z) {}
+    public record Home(double x, double y, double z) {}
 }
